uploading more examples

+++ examples/14_structs.sol
@@ -12,7 +12,6 @@ struct Todo {
 }
 
 contract Todos {
-
     // An array of 'Todo' structs
     Todo[] public todos;
 
@@ -34,19 +33,23 @@ contract Todos {
 
     // Solidity automatically created a getter for 'todos' so
     // you don't actually need this function.
-    function get(uint _index) public view returns (string memory text, bool completed) {
+    function get(uint256 _index)
+        public
+        view
+        returns (string memory text, bool completed)
+    {
         Todo storage todo = todos[_index];
         return (todo.text, todo.completed);
     }
 
     // update text
-    function updateText(uint _index, string calldata _text) public {
+    function updateText(uint256 _index, string calldata _text) public {
         Todo storage todo = todos[_index];
         todo.text = _text;
     }
 
     // update completed
-    function toggleCompleted(uint _index) public {
+    function toggleCompleted(uint256 _index) public {
         Todo storage todo = todos[_index];
         todo.completed = !todo.completed;
     }
