+++ examples/15_dataLocations.sol
@@ -4,7 +4,7 @@
 // memory - variable is in memory and it exists while a function is being called
 // calldata - special data location that contains function arguments
 
-// The storage, memory, and calldata labels are primarily used when you're working with variables within functions 
+// The storage, memory, and calldata labels are primarily used when you're working with variables within functions
 // or function parameters to specify where the data should be stored or how it should be treated within the function.
 
 // SPDX-License-Identifier: MIT
@@ -12,12 +12,12 @@ pragma solidity ^0.8.20;
 
 contract DataLocations {
     // considered to be stored in storage by default
-    uint[] public arr;
-    mapping(uint => address) map;
+    uint256[] public arr;
+    mapping(uint256 => address) map;
     struct MyStruct {
-        uint foo;
+        uint256 foo;
     }
-    mapping(uint => MyStruct) myStructs;
+    mapping(uint256 => MyStruct) myStructs;
 
     function f() public {
         // call _f with state variables
@@ -30,19 +30,19 @@ contract DataLocations {
     }
 
     function _f(
-        uint[] storage _arr,
-        mapping(uint => address) storage _map,
+        uint256[] storage _arr,
+        mapping(uint256 => address) storage _map,
         MyStruct storage _myStruct
     ) internal {
         // do something with storage variables
     }
 
     // You can return memory variables
-    function g(uint[] memory _arr) public returns (uint[] memory) {
+    function g(uint256[] memory _arr) public returns (uint256[] memory) {
         // do something with memory array
     }
 
-    function h(uint[] calldata _arr) external {
+    function h(uint256[] calldata _arr) external {
         // do something with calldata array
     }
 }
